refactor component folder to kebab-case
