+++ src/components/character-list/CharacterList.module.scss
