+++ src/components/character-list/generateKey.js
